requirements.txt
flask
pypdf
python-dotenv
openai>=1.0.0


